Created landing page and keypad game
I created a basic landing page with no CSS applied to it yet. Then I made a number guessing game in which you enter a 4 digit number with a keypad, and the number gets evaluated and compared to a random passcode. Then I created a small segment on the landing page to describe and link to the number guessing game page.