Temporary hosting of prank page
A friend of mine wanted to prank her grandma. She generated HTML, and asked me to  host it for her, since she has no web experience. This page was only used to prank her grandma, and for nothing else. It's not malicious. It's my intention to take it down once the prank is complete